refactor: Clean up of top-level tests (#1386)
