+++ packages/flame/test/anchor_test.dart
@@ -1,30 +1,31 @@
 import 'package:flame/extensions.dart';
 import 'package:flame/src/anchor.dart';
+import 'package:flame_test/flame_test.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('Anchor', () {
-    test('can parse to and from string', () {
+    test('parses to and from string', () {
       expect(Anchor.center.toString(), 'center');
       expect(Anchor.valueOf('topRight'), Anchor.topRight);
 
       expect(Anchor.values.length, 9);
-
       for (final value in Anchor.values) {
         final thereAndBack = Anchor.valueOf(value.toString());
         expect(thereAndBack, value);
       }
     });
 
-    test('can parse custom anchor', () {
-      expect(const Anchor(0.2, 0.2).toString(), 'Anchor(0.2, 0.2)');
-      expect(Anchor.valueOf('Anchor(0.2, 0.2)'), const Anchor(0.2, 0.2));
+    test('parses custom anchor', () {
+      const anchor = Anchor(0.2, 0.2);
+      expect(anchor.toString(), 'Anchor(0.2, 0.2)');
+      expect(Anchor.valueOf('Anchor(0.2, 0.2)'), anchor);
     });
 
     test('fail to parse invalid anchor', () {
       expect(
         () => Anchor.valueOf('foobar'),
-        throwsA(isA<AssertionError>()),
+        failsAssert(),
       );
     });
